+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Prepend clickable Jira issue link to Pull Request description when using `stud submit` [SCI-1]
 - Add `init` alias for `config:init` command [TPW-56]
 - Add interactive shell auto-completion setup prompt at the end of `config:init` command [TPW-56]
 - Automatically detect user's shell (bash or zsh) and provide installation instructions for shell completion [TPW-56]