+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.0] - 2025-11-20
+
 ### Added
 - Prepend clickable Jira issue link to Pull Request description when using `stud submit` [SCI-1]
 - Add `init` alias for `config:init` command [TPW-56]